Challenges/Web/fanpage/src/requirements.txt
Flask==3.0.3
waitress==2.1.2
PyJWT==2.9.0